// types/staffInterviewTypes.ts
export type EmployeeData = {
    id: string;
    area: string;
    branchCode: string;
    branchName: string;
    city: string;
    dateOfJoining: string;
    designation: string;
    employeeId: string;
    employeeName: string;
    name: string;
    postingOnNoticeBoard: string;
    province: string;
    region: string;
    staffName: string;
    visitDate: string;
    [key: string]: string;
};